Revert "Model/Folder: Use Integer instead of Float"
This reverts commit ca3931bd0d06cf500cff2443e4205f7c3211e777.

+++ app/src/main/java/com/nutomic/syncthingandroid/model/Folder.java
@@ -51,7 +51,7 @@ public class Folder {
     }
 
     public static class MinDiskFree {
-        public Integer value = 1;
+        public float value = 1;
         public String unit = "%";
     }
